Refactor Log4shell exploit code into reusable bits
